Added Very Basic Interaction + UI
Added a chest in spawn where you can "open" it. Right now it just logs to console but eventually we will trigger the animation and then modify the user inventory.

Also added an inventory  with just one key at the moment. To get the key simply press 'k'. Depending on this state the chest logs a different message when you press 'e' on it.